middleware: Check if sockets are open in update functions
I'm not sure if this will truly help, because I'm not sure if the state
can change inside the if statement. But it is possible that it cannot
and the crash inside updateX only because the socket is in a CLOSING
state and that means that .onClose didn't delete the socket from the
array yet.